```text feat: Add contact information to /sponsor-us page
This commit adds contact information to the /sponsor-us page. It includes the email address and phone numbers of the team members. The contact information is displayed in a visually appealing grid layout. This enhancement improves the user experience by providing an easy way for visitors to get in touch with the team.