feat: Update contact information on /sponsor-us page
This commit updates the contact information on the /sponsor-us page. It replaces the previous email address and phone numbers with new ones. The email address is now "rotaractgalgotias.web@gmail.com" and the phone numbers are "+91 95465 57824" for Areeb and "+91 97945 65358" for Divyanshu. This change improves the accuracy of the contact information and ensures that visitors can easily reach out to the team members.